requirements.txt
certifi==2019.6.16
chardet==3.0.4
Django==1.11.6
django-filter==2.2.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
idna==2.8
Markdown==3.1.1
mysqlclient==1.4.2.post1
Pillow==6.1.0
pkg-resources==0.0.0
Pygments==2.4.2
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.2
requests==2.22.0
urllib3==1.25.3
